Fix runningOnEC2 to work with all instances.
